Refactored files to prepare for Jest
