+++ components/sections/about/AboutSection.tsx
@@ -4,10 +4,10 @@ import Image from 'next/image'
 import { motion, useAnimation } from 'framer-motion'
 import { useInView } from 'react-intersection-observer'
 
-import Box from '../core/Box'
-import data from '../../data/data.json'
-import useBreakpoint from '../../shared/hooks/useBreakpoint'
-import me from '../../public/images/me.webp'
+import Box from "../../core/box/Box"
+import data from "../../../data/data.json"
+import useBreakpoint from "../../../shared/hooks/use-breakpoint"
+import me from "../../public/images/me.webp"
 
 const AboutSection = () => {
   const { t } = useTranslation()
