+++ components/sections/contact/ContactSection.tsx
@@ -2,8 +2,8 @@ import { useEffect } from 'react'
 import { useTranslation } from 'next-i18next'
 import { motion, useAnimation } from 'framer-motion'
 import { useInView } from 'react-intersection-observer'
-import Box from '../core/Box'
-import data from '../../data/data.json'
+import Box from '../../core/box/Box'
+import data from '../../../data/data.json'
 
 const ContactSection = () => {
   const { t } = useTranslation()
