+++ pages/index.tsx
@@ -3,22 +3,14 @@ import Head from 'next/head'
 import dynamic from 'next/dynamic'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
-const HeroSection = dynamic(() => import('../components/sections/HeroSection'))
-const AboutSection = dynamic(
-  () => import('../components/sections/AboutSection')
-)
-const SkillsSection = dynamic(
-  () => import('../components/sections/SkillsSection')
-)
-const WorksSection = dynamic(
-  () => import('../components/sections/WorksSection')
-)
-const ContactSection = dynamic(
-  () => import('../components/sections/ContactSection')
-)
+const HeroSection = dynamic(() => import('../components/sections/hero/HeroSection'))
+const AboutSection = dynamic(() => import('../components/sections/about/AboutSection'))
+const SkillsSection = dynamic(() => import('../components/sections/skills/SkillsSection'))
+const WorksSection = dynamic(() => import('../components/sections/works/WorksSection'))
+const ContactSection = dynamic(() => import('../components/sections/contact/ContactSection'))
 
-import Navigation from '../components/navigation/Navigation'
-import Footer from '../components/Footer'
+import Navigation from '../components/navigation/navigation-item/Navigation'
+import Footer from '../components/sections/footer/Footer'
 import data from '../data/data.json'
 
 const Home: NextPage = () => {
